Add UDP feature to tune the max packet size depending on the network protocol
Problem:
There is a hardcoded limit for the coap payload size.
Which is currently limitted to 1280 bytes.
However, UDP supports payload up to 64KB so this limitation is not
allowing to use all the benefits of UDP protocol.

Solution:
Adding an optional feature `udp` which if enabled will set the max payload size to
64000B, otherwise the default value that is currently set at 1280 will be used

https://github.com/martindisch/coap-lite/issues/25